Fix SK_File_GetSize for symlinks
SpecialK currently exhibits various problems when dealing with symlinks.
For example, if a configuration file is a symlink to some other file,
SpecialK will completely ignore that configuration file. This causes
problems for me because I use symlinks with some regularity.

It turns out the root cause is that SpecialK's function for checking the
size of a file does not work for symlinks because the Windows API call
that it uses (GetFileAttributesEx) always returns a file size of zero
if the specified target is a symlink. However, the _wstat64 function
works correctly for symlinks.

One possible solution would be to completely replace the use of
GetFileAttributesEx with _wstat64 instead. However, there are some
sutble differences between the two functions and I don't want to risk
introducing bugs by doing that. Instead, this change explicitly
checks if the target is a symlink and, if so, uses _wstat64 in that
case only. This cannot introduce any regressions because the code is
currently completely broken for symlinks, so this can only make it
better, not worse.